move containers to a separate directory
