+++ StRoot/StPicoD0EventMaker/StPicoD0EventMaker.cxx
@@ -8,14 +8,16 @@
 #include "StThreeVectorF.hh"
 #include "StLorentzVectorF.hh"
 #include "StPhysicalHelixD.hh"
-#include "../StPicoDstMaker/StPicoDst.h"
-#include "../StPicoDstMaker/StPicoDstMaker.h"
-#include "../StPicoDstMaker/StPicoEvent.h"
-#include "../StPicoDstMaker/StPicoTrack.h"
-#include "../StPicoDstMaker/StPicoBTofPidTraits.h"
-#include "StPicoD0Event.h"
+#include "StPicoDstMaker/StPicoDst.h"
+#include "StPicoDstMaker/StPicoDstMaker.h"
+#include "StPicoDstMaker/StPicoEvent.h"
+#include "StPicoDstMaker/StPicoTrack.h"
+#include "StPicoDstMaker/StPicoBTofPidTraits.h"
+
+#include "StPicoCharmContainers/StPicoD0Event.h"
+#include "StPicoCharmContainers/StKaonPion.h"
+
 #include "StPicoD0Hists.h"
-#include "StKaonPion.h"
 #include "StCuts.h"
 
 #include "StPicoD0EventMaker.h"
